Move widget to "Widget" folder
